af-delivery-data-platform/af-kb-chatbot-kb/pyproject.toml
[project]
name = "af-kb-chatbot-kb"
version = "0.0.1"
authors = [{ name = "datn@activefence.com" }]
requires-python = ">= 3.11"
dependencies = [
    "aiogoogle>=5.17.0",
    "aiolimiter>=1.2.1",
    "docling>=2.53.0",
    "google-api-python-client>=2.182.0",
    "google-auth>=2.40.3",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "langchain-openai>=0.3.33",
    "pinecone[asyncio]>=7.3.0",
    "ruff>=0.13.1",
    "tiktoken>=0.11.0",
    "tqdm>=4.67.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "databricks-dlt",
]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = [
    "tests",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/af-kb-chatbot-kb"]

[tool.basedpyright]
extraPaths = ["../shared"]

[tool.ruff]
line-length = 120
target-version = "py312"
src = [".", "../shared"]
exclude = [
    "__builtins__.pyi",
    "scratch"
]

[tool.ruff.lint]
select = [
    "E4",      # Import errors
    "E7",      # Statement errors  
    "E9",      # Runtime errors
    "F",       # Pyflakes
    "I",       # Import order
    "UP",      # pyupgrade
    "B",       # flake8-bugbear
    "C4",      # flake8-comprehensions
    "ICN",     # flake8-import-conventions
    "PIE",     # flake8-pie
    "PYI",     # flake8-pyi
    "RSE",     # flake8-raise
    "RUF",     # Ruff-specific rules
]
